tests/wpt/tests/css/css-gaps/tentative/parsing/rule-break-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: *-rule-break getComputedStyle()</title>
  <link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#break">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_computed_value("column-rule-break", "none");
    test_computed_value("column-rule-break", "spanning-item");
    test_computed_value("column-rule-break", "intersection");
    test_computed_value("row-rule-break", "none");
    test_computed_value("row-rule-break", "spanning-item");
    test_computed_value("row-rule-break", "intersection");
  </script>
</body>
</html>
